Simple UI HUD + Lantern Functionality
Added a simple HUD widget that uses C++ (for functionality) and blueprints (for basic layout). HUD displays Stamina used for sprinting (decays overtime as you use it and replenishes when not used), lantern fuel (decays over time at a rate providing 90 seconds of fuel), and text values for timers and sheep count that need to be updated. Need to add oil pickups for lantern refueling.